Add static_cast<int>() for the Qt size and index type when coming from std types

+++ src/track/serato/beatgrid.cpp
@@ -458,7 +458,7 @@ void SeratoBeatGrid::setBeats(BeatsPointer pBeats,
         }
     }
 
-    nonTerminalMarkers.reserve(markers.size());
+    nonTerminalMarkers.reserve(static_cast<int>(markers.size()));
     std::transform(markers.cbegin(),
             markers.cend(),
             std::back_inserter(nonTerminalMarkers),
